requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
-e git+https://github.com/jcanedo279/cointegridy.git@85e0f4dd30946a3e0525aa70834717e14e9081bc#egg=cointegridy
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fonttools==4.28.3
future==0.18.2
hurst==0.0.5
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.4
packaging==21.3
pandas==1.3.4
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycoingecko==2.2.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
statsmodels==0.13.1
threadpoolctl==3.0.0
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
treelib==1.6.1
urllib3==1.26.7
wcwidth==0.2.5
